Refactoring RuntimeItemStandard into its own file
