+++ src/vs/workbench/services/positronConsole/browser/classes/runtimeItemPendingInput.ts
@@ -1,11 +1,11 @@
 /*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2023-2025 Posit Software, PBC. All rights reserved.
+ *  Copyright (C) 2023-2026 Posit Software, PBC. All rights reserved.
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
 import { RuntimeCodeExecutionMode } from '../../../languageRuntime/common/languageRuntimeService.js';
 import { IConsoleCodeAttribution } from '../../common/positronConsoleCodeExecution.js';
-import { RuntimeItemStandard } from './runtimeItem.js';
+import { RuntimeItemStandard } from './runtimeItemStandard.js';
 
 /**
  * RuntimeItemPendingInput class.
