+++ src/vs/workbench/services/positronConsole/browser/classes/runtimeItemReconnected.ts
@@ -1,9 +1,9 @@
 /*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2023-2025 Posit Software, PBC. All rights reserved.
+ *  Copyright (C) 2023-2026 Posit Software, PBC. All rights reserved.
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { RuntimeItemStandard } from './runtimeItem.js';
+import { RuntimeItemStandard } from './runtimeItemStandard.js';
 
 /**
  * RuntimeItemReconnected class.
